+++ src/app/(main)/stories/new/StoryNewPageClient.tsx
@@ -1,0 +1,192 @@
+"use client";
+
+import { useState, Suspense, useEffect } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+import Image from "next/image";
+import { toast } from "react-hot-toast";
+import BookstoryText from "@/components/base-ui/BookStory/Editor/bookstory_text";
+import BookstoryChoosebook from "@/components/base-ui/BookStory/Editor/bookstory_choosebook";
+import BookSelectModal from "@/components/layout/BookSelectModal";
+import { useBookDetailQuery } from "@/hooks/queries/useBookQueries";
+import { useCreateBookStoryMutation } from "@/hooks/mutations/useStoryMutations";
+import { useAuthStore } from "@/store/useAuthStore";
+
+function StoryNewContent() {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const { isLoggedIn, isInitialized, openLoginModal } = useAuthStore();
+
+  useEffect(() => {
+    if (isInitialized && !isLoggedIn) {
+      toast.error("책 이야기 글 작성하기는 로그인이 필요한 서비스입니다.", { id: "story-create-auth-error-direct" });
+      router.replace("/stories");
+      setTimeout(() => {
+        openLoginModal();
+      }, 100);
+    }
+  }, [isInitialized, isLoggedIn, router, openLoginModal]);
+
+  const isbn = searchParams.get("isbn");
+  const [title, setTitle] = useState("");
+  const [detail, setDetail] = useState("");
+  const [isBookSelectModalOpen, setIsBookSelectModalOpen] = useState(false);
+
+  const { data: selectedBook } = useBookDetailQuery(isbn || "");
+  const createStoryMutation = useCreateBookStoryMutation();
+
+  const handleCancel = () => {
+    router.back();
+  };
+
+  const handleSubmit = (status: "PUBLISHED" | "DRAFT" = "PUBLISHED") => {
+    if (!selectedBook) {
+      toast.error("책을 선택해 주세요.");
+      return;
+    }
+    if (!title.trim()) {
+      toast.error("제목을 입력해 주세요.");
+      return;
+    }
+    if (status === "PUBLISHED" && !detail.trim()) {
+      toast.error("내용을 입력해 주세요.");
+      return;
+    }
+
+    createStoryMutation.mutate({
+      isbn: selectedBook.isbn,
+      title,
+      description: detail,
+      status,
+    }, {
+      onSuccess: () => {
+        if (status === "DRAFT") {
+          toast.success("임시저장되었습니다.");
+          router.push("/profile/mypage"); // 임시저장 시 마이페이지로 이동
+        } else {
+          toast.success("스토리가 등록되었습니다!");
+          router.push("/stories");
+        }
+      },
+      onError: (error) => {
+        console.error("스토리 등록 실패:", error);
+        toast.error(status === "DRAFT" ? "임시저장에 실패했습니다." : "스토리 등록에 실패했습니다. 다시 시도해 주세요.");
+      }
+    });
+  };
+
+  const handleBookSelect = (selectedIsbn: string) => {
+    router.push(`/stories/new?isbn=${selectedIsbn}`);
+  };
+
+  return (
+    <div className="relative mx-auto w-full max-w-[1400px] px-4">
+      {/* 책이야기 > 글 작성하기 - 모달 열리면 숨김 */}
+      {!isBookSelectModalOpen && (
+        <>
+          {/* 모바일: 전체 너비 선 */}
+          <div className="t:hidden w-screen -mx-4 border-b border-zinc-300">
+            <div className="px-4 h-[44px] flex gap-5 items-center">
+              <div className="body_1 text-Gray-3">전체</div>
+              <div className="relative w-[12px] h-[12px]">
+                <Image
+                  src="/triangle.svg"
+                  alt="next"
+                  fill
+                  className="object-contain"
+                />
+              </div>
+              <div className="body_1 text-Gray-7">글 작성하기</div>
+            </div>
+          </div>
+          {/* 태블릿/데스크탑: max-w 안에서 선 */}
+          <div className="hidden t:flex t:mt-6 h-[44px] gap-5 items-center border-b border-zinc-300">
+            <div className="d:subhead_4_1 text-Gray-3">전체</div>
+            <div className="relative w-[12px] h-[12px] d:w-[18px] d:h-[18px]">
+              <Image
+                src="/triangle.svg"
+                alt="next"
+                fill
+                className="object-contain"
+              />
+            </div>
+            <div className="d:subhead_4_1 text-Gray-7">글 작성하기</div>
+          </div>
+        </>
+      )}
+      {/* 메인 콘텐츠 영역 */}
+      <div className="py-6 px-2.5 t:px-10">
+        {/* 책 선택하기 박스 */}
+        <div className="flex justify-center mb-6">
+          {selectedBook ? (
+            <BookstoryChoosebook
+              bookUrl={selectedBook.imgUrl}
+              bookName={selectedBook.title}
+              author={selectedBook.author}
+              bookDetail={selectedBook.description}
+              onButtonClick={() => {
+                setIsBookSelectModalOpen(true);
+              }}
+            />
+          ) : (
+            <div className="flex w-full max-w-[1040px] p-5 justify-center items-center rounded-lg bg-White border-2 border-Subbrown-4">
+              <button
+                type="button"
+                onClick={() => setIsBookSelectModalOpen(true)}
+                className="flex px-6 py-2 justify-center items-center rounded-lg border border-primary-2 text-primary-2 body_1_2 bg-background cursor-pointer"
+              >
+                책 선택하기
+              </button>
+            </div>
+          )}
+        </div>
+
+        {/* 글 작성 영역 */}
+        <div className="flex justify-center">
+          <BookstoryText
+            title={title}
+            detail={detail}
+            onChangeTitle={setTitle}
+            onChangeDetail={setDetail}
+          />
+        </div>
+
+        {/* 하단 버튼 */}
+        <div className="flex justify-center">
+          <div className="flex w-full max-w-[1040px] justify-center t:justify-end gap-4 mt-6">
+            <button
+              type="button"
+              onClick={() => handleSubmit("DRAFT")}
+              disabled={createStoryMutation.isPending}
+              className="flex px-4 py-3 w-[132px] h-[44px] justify-center items-center rounded-lg border border-primary-1 text-primary-3 body_1_2 bg-background transition-colors hover:bg-primary-1 disabled:opacity-50"
+            >
+              {createStoryMutation.isPending ? "임시저장 중..." : "임시저장"}
+            </button>
+            <button
+              type="button"
+              onClick={() => handleSubmit("PUBLISHED")}
+              disabled={createStoryMutation.isPending}
+              className="flex px-4 py-3 w-[132px] h-[44px] justify-center items-center rounded-lg bg-primary-2 text-White body_1_2 hover:opacity-90 transition-opacity disabled:opacity-50"
+            >
+              {createStoryMutation.isPending ? "등록 중..." : "등록"}
+            </button>
+          </div>
+        </div>
+      </div>
+
+      {/* 책 선택 모달 */}
+      <BookSelectModal
+        isOpen={isBookSelectModalOpen}
+        onClose={() => setIsBookSelectModalOpen(false)}
+        onSelect={handleBookSelect}
+      />
+    </div>
+  );
+}
+
+export default function StoryNewPageClient() {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <StoryNewContent />
+    </Suspense>
+  );
+}
